Price Methods - Category Level

This function systematically runs the Price Method Productivity Output analysis for all species of a cateorgy.

PriceMethodOutput_Category(
  dat00,
  ii,
  category,
  category0,
  baseyr,
  maxyr,
  minyr,
  warnings_list = ls()
)

Arguments

dat00
Default dataset.

ii
Category number.

category
A character string. A unique string from the 'category0' column of the group being evaluated.

category0
A character string. The column where the category is defined.

baseyr
Numeric year (YYYY). The base year you are assessing the anaylsis with. Typically this is the earliest year in the data set, but it can be any year you choose.

maxyr
The maxium year to assess in the dataset.

minyr
The minium year to assess in the dataset.

warnings_list
A list where warnings are stored. If using this function in the PriceMethodOutput it will be inherited. If using outside of that function, put ls().