Reclassify ITIS species based off a list of higher taxonomic groupings

Reclassify ITIS species based off a list of higher taxonomic groupings

itis_reclassify(tsn, categories, missing_name)

Arguments

tsn
A vector of Taxonomic Serial Numbers to be evaluated.

categories
A list of the categories and associated TSN values.

missing_name
A string of what to call the missing value.

Value

df_out, tsn_indata

Examples

itis_reclassify(tsn = c(83677, 172746),
                categories = list("Finfish" = 914179, # Infraphylum  Gnathostomata
                                  "Shellfish" = c(82696, # Phylum  Arthropoda
                                                  69458)), # Phylum  Mollusca
                missing_name = "uncategorized")

#> $df_out
#>      TSN  category valid      rank              sciname
#> 1 172746   Finfish valid   species Scophthalmus aquosus
#> 2  83677 Shellfish valid subphylum            Crustacea
#> 3 172746 Shellfish valid   species Scophthalmus aquosus
#> 
#> $tsn_indata
#> $tsn_indata$`83677`
#>          name         rank     id
#> 1    Animalia      kingdom 202423
#> 2   Bilateria   subkingdom 914154
#> 3 Protostomia infrakingdom 914155
#> 4   Ecdysozoa  superphylum 914158
#> 5  Arthropoda       phylum  82696
#> 6   Crustacea    subphylum  83677
#> 
#> $tsn_indata$`172746`
#>                    name         rank     id
#> 1              Animalia      kingdom 202423
#> 2             Bilateria   subkingdom 914154
#> 3         Deuterostomia infrakingdom 914156
#> 4              Chordata       phylum 158852
#> 5            Vertebrata    subphylum 331030
#> 6         Gnathostomata  infraphylum 914179
#> 7        Actinopterygii   superclass 161061
#> 8             Teleostei        class 161105
#> 9       Acanthopterygii   superorder 166082
#> 10    Pleuronectiformes        order 172702
#> 11      Pleuronectoidei     suborder 172707
#> 12       Scophthalmidae       family 172713
#> 13         Scophthalmus        genus 172745
#> 14 Scophthalmus aquosus      species 172746
#> 
#>